properly report error when view query fails
CBLRouter.queryDesignDoc falsely reports success when
CBLView.queryWithOptions hits a SQLException and returns null. In that
case, queryWithOptions returns the current value of 'status', which is
always a successful one (see CBLRouter:L1463), resulting in a malformed
response.

This manifests itself in error logs like:

  [org.ektorp.ViewResult:L30] IllegalArgumentException: result must
  contain 'rows' field of array type

as Ektorp tries to make assertions about
the incorrectly-successful result.

This patch changes queryDesignDoc to return an INTERNAL_SERVER_ERROR
status in this case.