Add testLiveQuery() which uses a LiveQuery in the usual manner.  Currently fails due to apparent bug in code.
I didn't understand how the iOS TestCase(API_LiveQuery) works after glancing at it (maybe I should talk to Jens to try to figure it out), so for now I created a whole new testLiveQuery() which is the "normal" way to use a live query.  Left failingiOSPortTestLiveQuery() in for now, disabled.  Need to figure out if it's a valid test, and if so, re-enable it.