remove misleading "failed to rebuild view" logs
Log messages like

    W/Database: Failed to rebuild view Event/all: 500

are being printed every time a view is checked for re-indexing but is
already up to date. It looks like something's wrong, but it's a normal
state.

Setting the result to Status.OK suppresses these logs in the finally
block. The actual result (that reindexing was unnecessary) is still
printed at Debug level.