[Tests] TestAllDocuments now passes.
This brought in some code from the last snapshot that did
not previously get migrated to the current codebase. Now
that these changes are in place, and a couple of other
small errors I made in porting, the remaining failing
tests are failing from NUnit asserts, and not from
unhandled exceptions, by and large.