TestPusher stability improvement, additional logging.
Instead of UpdateActive calling Stopping synchronously,
when there are often other tasks in the worker queue,
add a task that calls stopped so that previous tasks
have a chance to run fist (e.g. SaveLastSequence).

Also, adds a log statement to the custom task scheduler
letting us know when a task is executed inline.