Changes behavior of new Docs.
Speculative branch that changes new Document
such that CurrentRevision returns an
UnsavedRevision instead of null. This
drives a number of other changes which
are accounted for in this commit.

Since this is a both a conceptual change--
Document currently is a collation of
SavedRevisions--and an API change--
CurrentRevision currently is typed
to return a SavedRevision or null,
this commit is made to a new branch.