Fix checkpoint save failure issue
CBSE-5429
Closes #1036 (Intermittent issue with saving checkpoint document)

The test is pretty obtuse, but basically it uses reflection to dispose of internal state at a key moment and checks to make sure that the next time the replicator is run it can recover to save the checkpoint to the local database.  Before it would fail with 10 != 20, but now it passes.