#1052 - covert datetimeoffset to string
because normally, JSON.net preserves the TimeZone offset in across the conversion. However, with Generics types, JSON.net decides on whether a string is DateTime, DateTimeOffset, or string based on custom serializer settings. By default, when a string parses to a DateTime or a DateTimeOffset, it chooses DateTime - which means that the offset information is dropped.