[CBSE-6069] Avoid a PerformWrite exception in .NET Native
Apparently, only in .NET Native compilation, disposing a BlockingCollection can race with Dispose to cause an internal NullReferenceException (as opposed to the ArgumentNullException that usually happens).  I added a catch for this but since it is native I worry if that is enough so I also added in a special lock so that "Take" will never be called on a disposed collection.  Fixes #1085