Fix some unneeded logic that was causing failures
Tests were failing on Xamarin iOS because the reachability component is more active and was causing an unwarranted reset in retry count.  Also there is a bug in TcpClient that causes a null reference exception when attempting to close / dispose an instance before it has an active connection, so catch and ignore it.