Revert just using the same message for two separate cases (db is closed and collection is not available)
After changed the error message and review the code, I prefer to keep the database is closed and the collection is deleted or not available separate for not open error. This will ensure the end user to clearly distinguish what is the actual issue behind the not open error when it occurs. This will also help us better to debug in case a CBSE is filed.