Fix UWP tests crashing when built with .NET Native (#1491)
This is tough to explain, but the lifetime management of the various interop objects was a bit sketchy and somehow, only in .NET Native, this caused push and/or pull filter functions defined by C# to become invalid (GC?  Memory move?) which meant that when LiteCore tried to call them it would get an unrecoverable native exception.  This PR creates a better hierarchy of objects that can have lifetimes tied to their parents.

There was also a few other issues found during this investigation dealing with tests.  The NonBlockingDatabaseOperation test was always throwing an exception and never actually verifying anything (the exception would kill the background thread and go unnoticed) and the "forum" test appears to not actually do anything and so it was removed.