CBL-4289: Don't skip waiting for pending conflict tasks
Previous code used to skip the conflict task check if there was no recorded error.  This seems like a meaningless check, so remove it and always check.  Also add the proper task into the conflict tasks list.  Previously the "continue with" tasks was being added, but the actual work task was being removed, which means the list would grow forever (full of completed tasks).