Add in .NET 6 Android specific workarounds
Very little functions as it should in .NET 6, and there are two specific large workarounds needed. First of all, none of this works without first either having a certificate issued by a CA trusted by default on Android, or adding in a trusted CA via network security config XML.  This creates a temporary contract that by the time the certificate reaches this callback, it is already trusted anyway...so a lot of this is just yak shaving until .NET 8.

1. When operating in normal mode (not "only self signed"), ask the underlying Java API is the certificate is trusted or not because X509Chain does not seem aware of everything that the Java API is aware of (specifically network security config trusted certificates).  This may or may not change in .NET 8, but for now at least it lines up with the fact that this is already trusted purely by the fact that it made it into this callback
2. When operating in self signed only mode, work around the non-functioning X509Chain by looking at the leaf cert directly (there is only supposed to be one entry anyway.  If it is self signed, that means there are no more certs in the chain)