Move from nuspec to csproj style packaging
This is largely the same process repeated several times.  Removing assembly info and letting the csproj generate it, assuming version injection via /p:Version="x.y.z" on the command line (default 0.0.0).  Making sure all dependencies are present in the csproj with proper versions (capped if necessary) and changing the LiteCore entries to be "None" with a package path so that they are included in the final nupkg.  This means that the "*.props" files will include the way to find the native library if necessary, though most of the platforms can simply find LiteCore by their proper placement in the runtimes folder.

Also Support.WinUI appears at this point to be redundant, and can be merged into Support.NetDesktop.  There is no significant difference between the two either in C# or in LiteCore.

The build and packaging scripts have been updated accordingly.