CBL-5304: Remove version specific Windows RID from nuget package (#1584)
Basically, when putting LiteCore into nuget put it into a folder named win-<arch> instead of win10-<arch>.  The reason I put it in win10-<arch> in the first place was to sort of have a few more hints that we weren't supporting Windows 7 or 8.  Nobody seems to be under that misconception anyway, now, and the toolchain is actually complaining about this so I will oblige.  It doesn't hurt the code any, as long as everything agrees on the location.

This will be kind of weird to test because PR validation is meaningless.  The post commit tests which run tests on the packaged library will show pretty quickly if anything is wrong.