Purge all compiler warnings (#1601)
* Purge all compiler warnings

The vast majority are simply the compiler being confused about the signature saying "not null" and then an immediate check if the value is null or not.  It errs on the safe side and issues warnings about a possible null value on some platforms but not others.  The compiler will let you know if you are potentially passing a null to a method that cannot accept it, and so the null checks are all less important now.

Also fill in missing doc comments and update all bad refs between doc pages (including changes to gen_defaults.py)