CBL-5000: Modify multiple replicator test to be less flaky
There is no reliable mechanism for capturing the max active count of multiple replicators from the passive side.  There are no listeners so the only choice is polling and sometimes polling can poll at the wrong time and give bad results.  Have to give up on this part of the test.