CBL-6329: QueryBuilder queries should dispose child objects
This is slightly strange in that some of the child objects are actually the same object as the parent, so check for that as well.  Each chain in a querybuilder call chain basically creates a new object that wraps the old one and appends additional information to itself, but the appended information was not being properly disposed.