Remove AddCollection API and finish converting all Configuration (#1725)
Now all Configuration objects are records.  ReplicatorConfiguration is slightly deviant, but presents publicly as immutable.  The endpoint wants to modify a few things on it, and the compiler cannot reason that this all happens at init time so I moved that information to its own class and had the endpoint modify that directly.  That way, when the configuration is copied the mutable slate is preserved, or set up again by the new target.  This means that something like `config with { Target = <new target] }` will behave as expected and a copy without changing the Target remains fast and simple.

The new pattern now is that every configuration has a constructor that contains all of the required properties as parameters, while also marking the required properties as such.  This means that a copy of a configuration object can be made with a modification to any property, including required ones.  If desired, the parameterless constructor can also be used and it is a compile error if the required properties are not all set.

* Rework the way IEndpoint works

This makes ReplicationConfiguration a lot easier to understand
Align CollectionConfiguration with the others in that the constructor still takes required items

* Make IndexConfiguration Expressions non nullable

* Rename CollectionConfigurations -> Collections