CBL-7410: Correct bug where partial indexes don't function in CE
The native struct layout differs depending on EE vs CE and that needs to be reflected.  The actual change here is in C4IndexTypes_defs.cs, and the rest is stuff that the bindings script picked up on its run.  I also took this chance to output some more info in the generate bindings script to warn me about which structs I should check when updating the bindings.