CBL-7410: Correct bug where partial indexes don't function in CE  (#1730)
* CBL-7410: Correct bug where partial indexes don't function in CE

The native struct layout differs depending on EE vs CE and that needs to be reflected.  The actual change here is in C4IndexTypes_defs.cs, and the rest is stuff that the bindings script picked up on its run.  I also took this chance to output some more info in the generate bindings script to warn me about which structs I should check when updating the bindings.

* Update LiteCore and potential database close hang fix

There were some test and bindings changes needed for 4.0.  Also there was a relatively high rate of hanging on the P2P tests when closing the database.  I think I have a way to poll to break out of that cycle now by introducing IAsyncStoppable

* Fix a redundant call to RemoveActiveStoppable

Add some doc comments and other cleanup to get rid of Release build warnings